Update HomePage.xaml.cs
Placed picker method on home page so user could select to search for properties for sale or either to let depending on selected choice.
Had difficulties with emulator testing app as was unable to get it to run due to issue mentioned below so was unable to test code edited to see if working or where to go next to complete objective.
Unable to test build as kept receiving error message form emulator - Couldn't connect debugger. You can see more details in Xamarin Diagnostic output and the full exception on logs.