Update PropertiesPage.xaml
Unable to run emulator due to error message - Couldn't connect debugger. You can see more details in Xamarin Diagnostic output and the full exception on logs.

Due to not being able to run emulator I was unable to check the UI to show that this is correct and can be read by the user.